Working on Warping
I have added a lot of the implementation for warping. Look into other activities taking over. See about warping in the MainActivity (threading) and then saving the results to memory. Then send over all saved image paths to the new activity to then display the data. Tonight work on scroll bar and passing in the original pictures to navigate through.