Added 3 new functions to WarpImage. Also got scaling done.
When importing, you must save then reload for the images to properly scale. Look into why the Uri isn't working for setImage. Also ask J for some help tomorrow in lab about getting the pixels and whatnot. Ask him to look over the loop and just to step through what needs to be saved and what needs to be changed